requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.13.1
elasticsearch==8.15.0
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==72.1.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.43.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
cohere==5.9.4